Cleanup of Swift 4 Codable for input field objects
